Moving sign out from profile to settings.
